进度条处理

+++ src/App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import {
   Button,
   Layout,
@@ -9,6 +9,7 @@ import {
   Typography,
   Upload,
   message,
+  Progress,
 } from "antd";
 import "./App.css";
 import {
@@ -25,36 +26,46 @@ const { Title } = Typography;
 const { Header, Content, Footer, Sider } = Layout;
 const { SubMenu } = Menu;
 
-const readFile = ( f ) => (new Promise((resolve, reject)=>{
+const readFile = (f) =>
+  new Promise((resolve, reject) => {
+    let fileReader = new FileReader();
 
-  let fileReader = new FileReader()
+    fileReader.onload = (evt) => {
+      resolve(evt.target.result);
+    };
 
-  fileReader.onload = (evt) => {
-    resolve(evt.target.result)
-  }
-
-  fileReader.readAsArrayBuffer(f)
-  
-}));
+    fileReader.readAsArrayBuffer(f);
+  });
 
 const App = () => {
   // const { result, error } = useWorker(createWorker, { a: 1 });
 
+  const [progress, setProgress] = useState(0);
+
   const startWorkerProcess = async (gzFile) => {
-    let fileData = await readFile(gzFile)
-    console.log(fileData)
+    let fileData = await readFile(gzFile);
+    console.log(fileData);
     // 创建 worker
-    let worker = new Worker(new URL("./lib/gzip_analyze.worker", import.meta.url));
+    setProgress(0.1);
+    let worker = new Worker(
+      new URL("./lib/gzip_analyze.worker", import.meta.url)
+    );
     worker.onmessage = (e) => {
-       //console.log('message from worker')
-       console.log(e)
-       worker.terminate()
-    }
-    worker.postMessage({type:'OPEN_FILE', payload:fileData}, [fileData])
-  }
+      //console.log('message from worker')
+      if (e.data.type === "INFLATE_RESULT") {
+        worker.terminate();
+        console.log(e.data);
+        setProgress(100);
+        setTimeout(()=>{setProgress(0)}, 1000)
+      } else if (e.data.type === "INFLATE_PROGRESS") {
+        setProgress(e.data.payload);
+      }
+    };
+    worker.postMessage({ type: "OPEN_FILE", payload: fileData }, [fileData]);
+  };
 
   const uploadProps = {
-    showUploadList:false,
+    showUploadList: false,
     beforeUpload: (file) => {
       console.log(file.type);
       if (file.type !== "application/x-gzip") {
@@ -68,8 +79,6 @@ const App = () => {
       startWorkerProcess(info.file);
     },
   };
-
-  
 
   return (
     <Layout>
@@ -89,13 +98,30 @@ const App = () => {
           dashed={true}
           style={{ borderColor: "rgba(255, 255, 255, 0.3)" }}
         ></Divider>
-        <Row justify="center">
-          <Space value="30">
-            <Upload {...uploadProps}>
-              <Button icon={<UploadOutlined />}>Open *.gz file</Button>
-            </Upload>
-          </Space>
-        </Row>
+        {progress === 0 ? (
+          <Row justify="center">
+            <Space value="30">
+              <Upload {...uploadProps}>
+                <Button icon={<UploadOutlined />}>Open *.gz file</Button>
+              </Upload>
+            </Space>
+          </Row>
+        ) : (
+          <Row justify="center">
+            <Space value="30">
+              <div style={{ width: 150 }}>
+                <Progress
+                  strokeColor={{
+                    "0%": "#108ee9",
+                    "100%": "#87d068",
+                  }}
+                  percent={progress}
+                  showInfo={false}
+                />
+              </div>
+            </Space>
+          </Row>
+        )}
         <Divider
           dashed={true}
           style={{ borderColor: "rgba(255, 255, 255, 0.3)" }}
